Refactor to extract standardize logic to module
