+++ app/models/simple_inline_text_annotation/generator.rb
@@ -1,5 +1,7 @@
 class SimpleInlineTextAnnotation
   class Generator
+    include DenotationStandardizer
+
     def initialize(source)
       @source = source
       @denotations = build_denotations(source[:denotations] || [])
@@ -7,7 +9,7 @@ class SimpleInlineTextAnnotation
 
     def generate
       text = @source[:text]
-      denotations = standardize_denotations(@denotations)
+      denotations = standardize(@denotations)
       config = @source[:config]
 
       annotated_text = annotate_text(text, denotations, config)
@@ -20,48 +22,6 @@ class SimpleInlineTextAnnotation
 
     def build_denotations(denotations)
       denotations.map { |d| Denotation.new(d[:span][:begin], d[:span][:end], d[:obj])}
-    end
-
-    # Standardize denotations by removing duplicates, nested, and boundary-crossing spans.
-    def standardize_denotations(denotations)
-      result = remove_duplicates_from(denotations)
-      result = remove_nesting_from(result)
-      remove_boundary_crossing_from(result)
-    end
-
-    def remove_duplicates_from(denotations)
-      denotations.uniq { |denotation| denotation.span }
-    end
-
-    def remove_nesting_from(denotations)
-      sorted_denotations = denotations.sort_by { |d| [d.begin_pos, -d.end_pos] }
-      result = []
-
-      sorted_denotations.each do |denotation|
-        unless result.any? { |r| r.begin_pos <= denotation.begin_pos && r.end_pos >= denotation.end_pos }
-          result << denotation
-        end
-      end
-
-      result
-    end
-
-    def remove_boundary_crossing_from(denotations)
-      denotations.reject do |denotation|
-        denotations.any? do |existing|
-          boundary_crossing?(denotation, existing)
-        end
-      end
-    end
-
-    def boundary_crossing?(denotation, existing)
-      is_start_of_denotation_span_between_existing_span =
-        existing.begin_pos < denotation.begin_pos && denotation.begin_pos < existing.end_pos && existing.end_pos < denotation.end_pos
-
-      is_end_of_denotation_span_between_existing_span =
-        denotation.begin_pos < existing.begin_pos && existing.begin_pos < denotation.end_pos && denotation.end_pos < existing.end_pos
-
-        is_start_of_denotation_span_between_existing_span || is_end_of_denotation_span_between_existing_span
     end
 
     def annotate_text(text, denotations, config)
